Older overlap.py file is renamed to ao_basis.py Edited significantly the function to construct AO bais: reduced the code, removed some unneeded stuff, fixed some methodological aspects and added some ideas to check. In general, we need more transferable (modular) code, more documentation, and more meaningfull function/variable names. Also, need to reduce the code size when possible - e.g. by avoiding redundancies. Updated the copyright notice in the two most edited files.
